Add ErrorList component
