+++ package.json
@@ -124,7 +124,7 @@
     "rc-dialog": "~8.3.0",
     "rc-drawer": "~4.1.0",
     "rc-dropdown": "~3.1.2",
-    "rc-field-form": "~1.10.0",
+    "rc-field-form": "~1.11.0",
     "rc-image": "~3.0.6",
     "rc-input-number": "~6.0.0",
     "rc-mentions": "~1.4.0",
